Push and Pull
Well, Push and pull has not gone as well as hoped. The "Attach to Actor" function sadly does not take into account the child actors collision. This inturn makes it very hard to do this as we have our child actors clipping through the environment. The plan tomorrow is to limit the turn rate of a player so they can only go forward and backwards. We will also take into account the froward vector of the push pull actor itself maybe and use that alone.